+++ internal/definitions/actions.lua
@@ -59,7 +59,7 @@ return {
     ArmTracks = 9,
     AutomationItem = 42197,
     AutoRenameTake = "_XENAKIOS_AUTORENAMETAKES",
-    BigItem = custom.select.innerBigItem,
+    BigItem = custom.innerBigItem,
     CleanProjectDirectory = 40098,
     ClearAllEnvelope = "_S&M_REMOVE_ALLENVS",
     ClearAllRecordArm = 40491,
@@ -140,8 +140,8 @@ return {
     DeleteTimeSignatureMarker = 40617,
     NextTimeSignatureMarker = 41821,
     PrevTimeSignatureMarker = 41820,
-    FirstItemStart = custom.move.firstItemStart,
-    FirstTrack = { custom.move.firstTrack, "ScrollToSelectedTracks" },
+    FirstItemStart = custom.firstItemStart,
+    FirstTrack = { custom.firstTrack, "ScrollToSelectedTracks" },
     GlueItemsIgnoringTimeSelection = 40362,
     FitByLoopingNoExtend = {
         "OnlySelectItemsCrossingTimeAndTrackSelection",
@@ -187,13 +187,13 @@ return {
     InvertVoicingDownwards = { 40910, midiCommand = true, prefixRepetitionCount = true },
     InvertVoicingUpwards = { 40909, midiCommand = true, prefixRepetitionCount = true },
     ApplyFxToItem = 40209,
-    Item = custom.select.innerItem,
+    Item = custom.innerItem,
     ItemNormalize = 40108,
     ItemSplitSelRight = "_SWS_AWSPLITXFADELEFT",
     JoinNotes = { "SelectNotes", "JoinSelectedNotes" },
     JoinSelectedNotes = { 40456, midiCommand = true },
-    LastItemEnd = custom.move.lastItemEnd,
-    LastTrack = { custom.move.lastTrack, "ScrollToSelectedTracks" },
+    LastItemEnd = custom.lastItemEnd,
+    LastTrack = { custom.lastTrack, "ScrollToSelectedTracks" },
     Left10Pix = { "_XENAKIOS_MOVECUR10PIX_LEFT", prefixRepetitionCount = true },
     Left40Pix = { "Left10Pix", repetitions = 4, prefixRepetitionCount = true },
     LeftGridDivision = { 40646, prefixRepetitionCount = true },
@@ -270,13 +270,13 @@ return {
     Next4Measures = { "NextMeasure", repetitions = 4, prefixRepetitionCount = true },
     Next5Track = { "NextTrack", repetitions = 5, prefixRepetitionCount = true },
     NextBeat = 40841,
-    NextBigItemEnd = { custom.move.nextBigItemEnd, prefixRepetitionCount = true },
-    NextBigItemStart = { custom.move.nextBigItemStart, prefixRepetitionCount = true },
+    NextBigItemEnd = { custom.nextBigItemEnd, prefixRepetitionCount = true },
+    NextBigItemStart = { custom.nextBigItemStart, prefixRepetitionCount = true },
     NextEnvelope = { 41864, prefixRepetitionCount = true },
     NextEnvelopePoint = { "_SWS_BRMOVEEDITTONEXTENV", prefixRepetitionCount = true },
     NextFolderNear = { "_SWS_SELNEARESTNEXTFOLDER", "ScrollToSelectedTracks", prefixRepetitionCount = true },
-    NextItemEnd = { custom.move.nextItemEnd, prefixRepetitionCount = true },
-    NextItemStart = { custom.move.nextItemStart, prefixRepetitionCount = true },
+    NextItemEnd = { custom.nextItemEnd, prefixRepetitionCount = true },
+    NextItemStart = { custom.nextItemStart, prefixRepetitionCount = true },
     NextMarker = { 40173, prefixRepetitionCount = true },
     NextMeasure = { 40839, prefixRepetitionCount = true },
     NextNoteEnd = { "SelectNextNote", "EventSelectionEnd", prefixRepetitionCount = true },
@@ -370,11 +370,11 @@ return {
     Prev4Measures = { "PrevMeasure", repetitions = 4, prefixRepetitionCount = true },
     Prev5Track = { "PrevTrack", repetitions = 5, prefixRepetitionCount = true },
     PrevBeat = { 40842, prefixRepetitionCount = true },
-    PrevBigItemStart = { custom.move.prevBigItemStart, prefixRepetitionCount = true },
+    PrevBigItemStart = { custom.prevBigItemStart, prefixRepetitionCount = true },
     PrevEnvelope = { 41863, prefixRepetitionCount = true },
     PrevEnvelopePoint = { "_SWS_BRMOVEEDITTOPREVENV", prefixRepetitionCount = true },
     PrevFolderNear = { "_SWS_SELNEARESTPREVFOLDER", "ScrollToSelectedTracks", prefixRepetitionCount = true },
-    PrevItemStart = { custom.move.prevItemStart, prefixRepetitionCount = true },
+    PrevItemStart = { custom.prevItemStart, prefixRepetitionCount = true },
     PrevMarker = { 40172, prefixRepetitionCount = true },
     PrevMeasure = { 40840, prefixRepetitionCount = true },
     PrevNoteEnd = { "SelectPrevNote", "EventSelectionEnd", prefixRepetitionCount = true },
@@ -413,9 +413,9 @@ return {
     },
     NextTransientInSelectedItemsMinusFadeTime = "_XENAKIOS_MOVECURNEXT_TRANSMINUSFADE",
     PrevTransientInSelectedItemsMinusFadeTime = "_XENAKIOS_MOVECURPREV_TRANSMINUSFADE",
-    ProjectEnd = custom.move.projectEnd,
-    ProjectStart = custom.move.projectStart,
-    ProjectTimeline = custom.select.innerProjectTimeline,
+    ProjectEnd = custom.projectEnd,
+    ProjectStart = custom.projectStart,
+    ProjectTimeline = custom.innerProjectTimeline,
     Quantize = { 40009, midiCommand = true },
     RecallMark = { lib.marks.recall, registerAction = true },
     RecordConditional = { "Stop", "ToggleRecordConditional" },
@@ -429,7 +429,7 @@ return {
         setTimeSelection = true
     },
     Redo = { 40030, prefixRepetitionCount = true },
-    Region = custom.select.innerRegion,
+    Region = custom.innerRegion,
     RegionSelectItems = 40717,
     RemoveMarker = 40613,
     RemoveRegion = 40615,
@@ -505,7 +505,7 @@ return {
     SelectNoteClosestToEditCursor = { 40426, midiCommand = true },
     SelectNotes = "SelectNotesStartingInTimeSelection",
     SelectNotesStartingInTimeSelection = { 40877, midiCommand = true },
-    SelectOnlyCurrentTrack = custom.select.onlyCurrentTrack,
+    SelectOnlyCurrentTrack = custom.onlyCurrentTrack,
     SelectOnlyFoldersChildren = "_SWS_SELCHILDREN",
     SelectPrevNote = { 40414, midiCommand = true },
     SelectPrevNoteSamePitch = { 40427, midiCommand = true },
@@ -520,7 +520,7 @@ return {
     SetRippleEditAllTracks = 40311,
     SetRippleEditOff = 40309,
     SetRippleEditPerTrack = 40310,
-    FirstTrackWithItem = custom.move.firstTrackWithItem,
+    FirstTrackWithItem = custom.firstTrackWithItem,
     DuplicateTimeline = {
         "SaveTrackSelection",
         "SelectAllTracks",
@@ -605,7 +605,7 @@ return {
     ShowPrevFx = { "_S&M_WNONLY1", prefixRepetitionCount = true },
     ShowProjectSettings = 40021,
     ShowTrackFreezeDetails = 41654,
-    SnappedPosition = custom.move.snap,
+    SnappedPosition = custom.snap,
     AddAndNameSnapshot = "_SWSSNAPSHOT_NEWEDIT",
     CopyCurrentSnapshot = "_SWSSNAPSHOT_COPY",
     PasteSnapshot = "_SWSSNAPSHOT_PASTE",
@@ -703,7 +703,7 @@ return {
     TrackToggleSolo = 40281,
     TrackToggleSoloDefeat = 41199,
     TrackToggleMute = 40280,
-    TrackWithNumber = custom.move.trackWithNumber,
+    TrackWithNumber = custom.trackWithNumber,
     TrimItemLeftEdge = 41305,
     TrimItemRightEdge = 41311,
     TrimSelectedNoteLeftEdgeToEditCursor = 40790,
